mcp_server/requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-mgmt-netapp
python-dotenv
httpx
